Add short output into AudioRecorder
